added test for changing the pos_tol / gap_tol / move_tol on restart
Note: when the parameter was None on the first run (in particular for pos_tol),
it is not required that the results are exactly the same afterwards. This is
because there is no Control to watch over the given convergence parameter. If
this is not the case, every ConvergenceControl always has to run.